Added graphical bits of the console system.
Added a few nice features: tab-autocomplete, suggestions, chat messages are displayed seperate from console commands
The console implementation follows our usual MVC style. Currently the objects are instantiated with the GameSession--eventually (when autofac moves to the top of the gameloop) we can make the console a singleton that persists over various GameSessions.
Changes to ISpriteBatch and IFont to support better text drawing. The LineSpacing property on IFont is useful for writing vertical lines of text.
Added functionality to the Keyboard class (now called XnaKeyboard and in Frenetic.UserInput namespace):
	The Keyboard can be locked (so no keys will register as pressed until Unlock() is called)
	SaveState() takes a snapshot of the current keyboard input. When WasKeyDown(Keys key) is then called, it will return true based on this saved state
	AlphaNumericKeys is an IEnumerable class useful for reading user text input

Made the level pieces a little harder to get stuck in, :P
I've just copied an FPS component from somewhere... This can serve as a model when we implement our own.