Basic console structure is now working with a mediator and delegates to change properties.
Some changes to GameConsole (now does better parsing of commands and parameters).
Dropped the Command class since I wasn't really using it as anything but a string wrapper.
Mediator is no longer a weak dictionary of delegates (this caused some problems, and it wasn't very easy for clients of Mediator to use it correctly). I think a better approach will be to have SingletonScope objects that are as long lived as the Console.
Added PlayerSettings and PhysicsSettings.
MediatorPlayerSettingsController now registers the command "PlayerName" with the mediator (this doesn't do anything yet though).