Console now reads numbers. Refactored PlayerView to use an Autofac factory. Added PlayerColor to possible console variables (and it works!).
The player texture is now stored on the PlayerSettings object. This is currently a singleton, but could be per player eventually. Hopefully this turns out to be a good architectural decision, :P.
The MediatorXXXControllers currently do the input validation. Not sure if this is the best place to do it...
PlayerView now uses PlayerSettings to determine how to render the player--so the player color can be changed from the console.
I had to add an extension method to Keys to read in numbers (at least, this seemed like the most elegant solution...)