Moved the console to be at the Outer container scope.  PlayerSettings now per-player instance.
There's still some refactoring needed for the way we use Autofac. Plus, the inner containers need to be disposed when you exit a game session back to the main menu.
All textures used by Players should now be loaded into an ITextureBank<PlayerTextures> object. Hopefully this system can be used generally (a LevelTextures TextureBank, etc.)
GameSessionController no longer creates the PlayerView for the local Player (the whole GameSessionController class is still too messy...)
On the other hand, GameSessionFactory is looking a little cleaner...
I'm a bit worried that the PlayerView doing essentially a ContentManager.Load<> on each Generate might be slow (reflection, etc.), but this can always be cached later and is the simplest way for now.
I wrapped the ContentManager with an IContentManager interface. Other classes should be changed to use this interface too...