Client now sends new chat messages to server until they are received by the server. Server only adds these messages once.
ChatLog now stores a new type: ChatMessage. This keeps the client name associated with the message, and a snap when it was created. This allows for much easier diffing and is probably a better solution anyway...
Because of the above change, Log is now generic (currently used as Log<string> and Log<ChatMessage>)
ChatLogDiffer now done in memory. This works well as long as the ChatLogSender doesn't try to modify the ChatMessages (which it shouldn't).