Player state and PlayerSettings now use new network architecture (but still doesn't diff to only send changed state).
NetworkPlayerProcessor now handles updating the Player and PlayerSettings from incoming messages.
ChatLogSender (soon to be renamed) and ClientInputSender now also send the Player state and PlayerSettings (still needs to do diffs)


Started changing the scope of some of the objects... There's now a LocalClient which is essentially just a copy of Client but that has application lifetime.
Currently I'm not very happy with how the Player property on the LocalClient needs to be set once the physics simulator is instantiated... Maybe it's better to make the physics simulator and the player also as long lived as the LocalClient?
GameSessionController is really getting a lot cleaner now (finally). I have some ideas for refactoring the new player part that should further improve GameSessionController.