Implemented generic [Tweakable] attribute for anything that should be changeable from the console (and is therefore a setting that is persisted between sessions)
A big change is that IMediator is now responsible for all the string conversion/parsing (so it's only done ONCE per type). There's lots of Generics bits to get it to work, but really adding a new type is easy.
Another big change is the introduction of reflection, so that PropertyInfo objects are now registered with the Mediator (plus the specific instance on which the property should be called)

NOTE: Files that need to be scanned for [Tweakable] attribute need to be loaded with the new class, TweakablePropertiesLoader.

There are now two types of PlayerSettings:  LocalPlayerSettings and NetworkPlayerSettings (this allows for easier object lifetime handling).

Unrelated... Moved all classes related to the Player into their own namespace (this has the nice benefit of giving prettier [Tweakable] names).