Implemented basic automatic Settings save and load.
Settings are loaded when the game starts, and saved whenever the the game is stopped normally.

Had to change the formatting of strings for Mediator.Get() to match the way parameters are fed to Mediator.Set().
Game exit now handled by a service called Quitter.
Removed the border (looks better + now have to quit through game menu/console)