Client joining now handled at the NetworkSession level, with events raised to be handled by the ClientStateTracker. This paves the way for a more generic delta compressed unreliable data network architecture (i hope!)
ClientStateTracker now handles all new Client configuration (PlayerViews, etc.) It does some of this with a new IClientFactory class (maybe slightly miss named?)
IncomingMessageQueue now has a HasAvailable method that makes for a slightly nicer way of reading messages off the queues.
The good news is that GameSessionController is now almost trivially simple (after getting really ugly for a while, with a huge constructor)
Some of the ServerNetworkSession work now delegated to a ServerMessageSender class.