Log output now also gets written to the in-game console.
Added a custom log4net Appender.
NOTE: It's a bit of a hack, but I had to use property injection on GameConsoleAppender to give it an instance of the CommandConsole Log.
This currently happens as soon as the CommandConsole gets instantiated (using the Autofac OnActivating event, which is SUPER useful!).

TODO: Make more use of the OnActivating event for object initialization.