Mediator now logs all messages through the logger interface.
This seems to be working quite nicely, but I'm a bit nervous about it. Not sure if this is really "proper". It's also not covered by unit tests.
On the other hand, it should be pretty visible if something goes wrong, and generally I've given up on trying to get test coverage of drawing code (it hasn't helped much up until now...).