MORE Network refactoring. Packet size about 1/3rd of what it was when this all started.
The packet contents is pretty optimized now (there is still some room for improvement, but at the cost of clarity).
The next step is to replace XmlSerialization. This will give HUGE benefits, but I'd rather delay it as late as possible. It's sometimes useful to be able to print out the serialized xml and see what's going on.

Introduced an IDiffable<> Interface:  This is implemented by anything that only needs to be sent when changed.
It was surprisingly easy to modify existing structures to work with this. The following are now only sent when changed:
	ChatLog, PlayerSettings, and Shots

It's now easier to add loggers to classes (using the ILoggerFactory). This plays nicer with Autofac.
PROBLEM: I'm not sure if I'm controlling logging levels correctly. Doing Debug output on every tick makes the game unplayably slow. Should I be enabling/disabling particular modules somehow?