Initial client-side prediction implementation
Based on the very simple idea of looking at the last 2 received positions and extrapolating for a 3rd "predicted" position. This is then used to set the object velocity on the PhysicsComponent.

Split Player into a LocalPlayer and NetworkPlayer. This polymorphism abstracts away some switching logic and the need for the PlayerType enum.