Refactored PlayerInput handling to make it (eventually) easier to handle respawns on the Server side.
PlayerInput now works in a similar way to PlayerState.
Introduced a PlayerRespawner class that will find a valid respawn position, update the Player and make it alive again.
PlayerUpdater will be responsible for Respawning Players via the PlayerRespawner class (on the Server side, as a response to a PendingShot).

Moved Level module to withing Gameplay module.

NEXT STEP: Add a "PlayerState" to the PlayerInput, and on the server side, only update the player if the states are in sync...