After several false starts, player respawning now seems to work...
The crux of the problems I had:  Clients are the authority on Player position, and the Server is the authority on Player Status (Alive or Dead at the moment).
The eventual solution was to do a respawn on the Client side by changing the Position (which the Client controls), and then requesting a Status change on the Server through the PendingStatus property on PlayerInput.

PlayerUpdater on the Server now checks for PendingStatus and changes the Status of the Server side Players accordingly.
The PlayerState class on the Client side checks to see if the incoming Status matches the PendingStatus, and if it does, it clears the PendingStatus.