Made commands executed through the Mediator do implicit type conversion where possible.
Hooked up the "Connect" command so you can connect to a specific ip/port. Not sure how this will work if you run it again from within a game!

Also unified the way Properties and Methods are handled, and as a result the code became much shorter/simpler.

I switched from Dictionaries of delegates, to Dictionaries of PropertyInfo and MethodInfo type. This seems to make things simpler down the line.
Rather than storing the Setter and Getter now with specific type info, the conversion is done at call time just like for methods. This means they can all call the generic ConvertTo<Type> method in the same way.