Add a BubbleFont class that's used to write the damage done in animated text.
Added a HealthChanged event to the Player. This event is used to write BubbleText any time player health changes.
Also had to create a PlayerList container class that holds a List<IPlayer> and fires a PlayerAdded event every time a player is added. The PlayerView uses this event to register for events on new players...
IView.Generate() now includes elapsedSeconds parameter.