Changed interface of IEffect to rather be a collection of effects. Trigger now takes a parameter of EffectType (and needs properties like Position set first).
This is consistent with Mercury's recommended approach of Emitters as Service rather than per object Emitters. The idea is to have one Emitter instance per effect type. Will see if it stays practical as number of effects increases (not very polymorphic right now...).

Rockets now have a trail and then an explosion. Yay!

Next step is to consolidate this new approach with the old Railgun approach and check that it all still works (by adding weapon switching?).